genai-assistant-main/requirements.txt
streamlit==1.46.1
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.4.1
